review 3개 insert 완료

+++ src/dao/air/AirReviewDao.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public interface AirReviewDao {
 	
-	void insert(String writer, String airContent, int airHoroscope, int airNo);
+	void insert(String writer, String airContent, double airHoroscope, int airNo);
 	
 	List<Integer> check_member(int memberNo);
 	double avg_horoscope(int airNo);
